backend/requirements_v2.txt
# Production dependencies for OptionsAI Copilot v2

# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Async and Redis
redis[hiredis]==5.0.1
aiohttp==3.9.1
asyncio-mqtt==0.16.1

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.13.0

# Market Data APIs
alpha-vantage==2.3.1
yfinance==0.2.28

# AI/ML
transformers==4.36.0
torch==2.1.0
tokenizers==0.15.0

# Data processing
pandas==2.1.4
numpy==1.25.2

# Background jobs
celery==5.3.4
redis==5.0.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0